requirements.txt
aiolimiter==1.0.0
anyio==3.6.2
APScheduler==3.10.1
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
cryptography==40.0.2
filelock==3.12.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.0
httpx==0.24.0
hyperframe==6.0.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
lit==16.0.3
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
Pillow==9.5.0
pycparser==2.21
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
regex==2023.5.5
requests==2.30.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
sympy==1.12
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.2
